--- a/math/integer/select.go
+++ b/math/integer/select.go
@@ -1,0 +1,15 @@
+package integer
+
+func Min(a, b int) int {
+	if a < b {
+		return a
+	}
+	return b
+}
+
+func Max(a, b int) int {
+	if a > b {
+		return a
+	}
+	return b
+}