… clone/clone_test.go
… convolution/convolution_test.go
… effect/effect.go
… effect/effect_test.go
… math/integer/helpers.go
… math/integer/helpers_test.go
… transform/resize.go
… transform/resize_test.go
… transform/rotate.go
… transform/rotate_test.go
… util/util.go