Improved performance of convolution operation. 
Refactored function for readability.
Rewrote tests so that they are more comprehensive.
Added new Pad method for clone package (extracted function from convolution optimisation).

Benchmarks:
```
benchmark                 old ns/op       new ns/op       delta
BenchmarkConvolve3-8      40166558        36045396        -10.26%
BenchmarkConvolve8-8      210586006       172509259       -18.08%
BenchmarkConvolve32-8     3118727037      2643871731      -15.23%
BenchmarkConvolve64-8     12936959358     10726478692     -17.09%

benchmark                 old allocs     new allocs     delta
BenchmarkConvolve3-8      6              8              +33.33%
BenchmarkConvolve8-8      6              8              +33.33%
BenchmarkConvolve32-8     6              8              +33.33%
BenchmarkConvolve64-8     7              8              +14.29%

benchmark                 old bytes     new bytes     delta
BenchmarkConvolve3-8      8389023       8413602       +0.29%
BenchmarkConvolve8-8      8388872       8462675       +0.88%
BenchmarkConvolve32-8     8388888       8659288       +3.22%
BenchmarkConvolve64-8     8388928       8929584       +6.44%
```